requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.8
DicomAnonymizer==0.0.1
idna==3.3
importlib-metadata==4.8.2
multivolumefile==0.2.3
numpy==1.21.4
pandas==1.4.2
py7zr==0.17.0
pybcj==0.5.0
pycryptodomex==3.12.0
pydicom==2.2.2
pyppmd==0.17.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzstd==0.15.0
requests==2.26.0
six==1.16.0
texttable==1.6.4
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
dicom2nifti==2.3.0
Pillow==9.0.0
SimpleITK==2.1.1.2